requirements.txt
pandas
numpy
scikit-learn
xgboost
fastapi
uvicorn
pyyaml
